Conflicts:
	musica-02-web/src/main/java/co/edu/uniandes/csw/musica/dtos/ClienteDetailDTO.java

+++ musica-02-logic/src/main/java/co/edu/uniandes/csw/musica/ejbs/EntradaLogic.java
@@ -34,7 +34,7 @@ public class EntradaLogic {
         return persistence.findByCliente(usuario);
     }
     
-    public List<EntradaEntity> getByFuncion(Integer id) {
+    public List<EntradaEntity> getByFuncion(Long id) {
         return persistence.findByFuncion(id);
     }
